Harite v3.1
- Refinement of almost all parts, except for cap top. The joysticks feel
tighter with less wobbly. It is definitely worth reprinting everything.
- Tolerance around base reduced from 0.4mm to 0.2mm. I found the
joysticks to be very sensitive to even a small offset from the centre,
they would crunch in one direction and the the actuation force would be
higher. There is still some wiggle room but this makes it a lot easier
to find the best position.
- Simplified joystick recess to reduce friction surface area.
- Added holes on top to push reset and see charging led on MCU.
Comfortably fits a 2.3mm diameter toothpick.
- Added slots to the Cirque and thumb cluster base to help hide the FFC
cable.
- Moved thumb cluster out a few mm and rotated inward slightly since I
was finding my index finger bumping into it sometimes.